refactor: react memo & money provider refactoring
