+++ src/App.jsx
@@ -3,17 +3,12 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import GlobalStyle from 'common/globalStyle';
 import { composeProvider } from 'common/utils';
 import Layout from 'components/Layout';
-import { ItemProvider } from 'context/ItemProvider';
 import { LogProvider } from 'context/LogProvider';
 import { MoneyProvider } from 'context/MoneyProvider';
 import VendingMachine from 'pages/VendingMachine';
 import Wallet from 'pages/Wallet';
 
-const ContextProvider = composeProvider([
-  LogProvider,
-  MoneyProvider,
-  ItemProvider,
-]);
+const ContextProvider = composeProvider([LogProvider, MoneyProvider]);
 
 const App = () => (
   <div className='App'>
