+++ src/components/VendingMachine/VMItem.jsx
@@ -1,9 +1,8 @@
-import { useCallback, useContext } from 'react';
+import { memo, useCallback } from 'react';
 import styled from 'styled-components';
 
 import { parseMoneyFormat } from 'common/utils';
 import COLORS from 'constants/colors';
-import { MoneyContext } from 'context/MoneyProvider';
 
 const defaultStyle = `
   border: 3px solid ${COLORS.GREY};
@@ -23,35 +22,32 @@ const vmItemStyleMap = {
   default: defaultStyle,
 };
 
-const VMItem = ({ item: { id, name, amount, count }, onClickActiveItem }) => {
-  const { inputMoney, buyVMItem } = useContext(MoneyContext);
-  const isSoldOut = count === 0;
-  const isActive = inputMoney >= amount && !isSoldOut;
+const VMItem = memo(
+  ({ item: { id, name, amount }, onClickActiveItem, isSoldOut, isActive }) => {
+    const getStatus = useCallback(() => {
+      if (isSoldOut) return 'soldout';
+      return isActive ? 'active' : 'default';
+    }, [isSoldOut, isActive]);
 
-  const getStatus = useCallback(() => {
-    if (isSoldOut) return 'soldout';
-    return isActive ? 'active' : 'default';
-  }, [isSoldOut, isActive]);
+    const handleClickActiveItem = () => {
+      if (!isActive) return;
+      onClickActiveItem({ id, amount, name });
+    };
 
-  const handleClickActiveItem = () => {
-    if (!isActive) return;
-    onClickActiveItem(id);
-    buyVMItem(amount, name);
-  };
-
-  return (
-    <li>
-      <ItemWrapper>
-        <ItemNameBox status={getStatus()} onClick={handleClickActiveItem}>
-          <span>{name}</span>
-        </ItemNameBox>
-        <ItemPriceBox>
-          <span>{parseMoneyFormat(amount)}</span>
-        </ItemPriceBox>
-      </ItemWrapper>
-    </li>
-  );
-};
+    return (
+      <li>
+        <ItemWrapper>
+          <ItemNameBox status={getStatus()} onClick={handleClickActiveItem}>
+            <span>{name}</span>
+          </ItemNameBox>
+          <ItemPriceBox>
+            <span>{parseMoneyFormat(amount)}</span>
+          </ItemPriceBox>
+        </ItemWrapper>
+      </li>
+    );
+  }
+);
 
 const ItemWrapper = styled.div`
   display: grid;
